add automatic cloud login route
