+++ .changeset/vast-corners-see.md
@@ -1,0 +1,5 @@
+---
+"@medusajs/dashboard": patch
+---
+
+add automatic cloud login endpoint
