+++ packages/admin/dashboard/src/routes/cloud/login/index.ts
@@ -1,0 +1,3 @@
+export { CloudLogin as Component } from "./login"
+
+export const CLOUD_AUTH_PROVIDER = "cloud"
